Rename generateAhoXml to generateLegacyAhoXml to better reflect how this is used
