+++ src/lib/generateLegacyAhoXml.ts
